Create index.js in components :ContentArea, JSONToHTML; move ContentArea to "CMS" section App.js
