+++ frontend/src/App.js
@@ -3,11 +3,10 @@ import React, { Component } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
 import LandingPage from './LandingPage';
-import Login from './Login';
 import Navbar from './Navbar';
 import StateSetup from './_StateSetup';
 
-import ContentArea from './ContentArea/ContentArea.js';
+import ContentArea from './ContentArea';
 
 // This is just a mock showing a simple react component included.
 // Viktor may remove this file when he does the routing, but you can
@@ -31,9 +30,9 @@ class App extends Component {
 
   render() {
     const { isLoggedIn, username } = this.state;
-
-    return (
-
+	
+   return (
+		
       <Router>
         <React.Fragment>
           <StateSetup
@@ -48,15 +47,16 @@ class App extends Component {
             <Route path="/pages" render={() => <div>Pages component</div>} />
             <Route path="/users" render={() => <div>Users component</div>} />
             <Route path="/assets" render={() => <div>Assets component</div>} />
-            <Route path="/cms" render={() => <div>CMS component</div>} />
-            <Route path="/login" component={Login} />
+            <Route path="/cms"  component={ContentArea} />         
+            <Route path="/login" render={() => <div>Login component</div>} />
             <Route path="/logout" render={() => <div>Logout component</div>} />
           </Switch>
+          
         </React.Fragment>
         
       </Router>
 
-    );
+    );     
   }
 }
