+++ frontend/src/ContentArea/index.js
@@ -1,0 +1,4 @@
+// each container uses an index.js file to expose it's reuseable parts
+import ContentArea from './ContentArea';
+
+export default ContentArea;
