+++ frontend/src/JSONHandle/index.js
@@ -1,0 +1,3 @@
+import JSONToHTML from './JSONToHTML';
+
+export {JSONToHTML};
